requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
blessed==1.20.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.4
cssselect2==0.7.0
daphne==4.1.0
Django==5.0.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-picklefield==3.1
django-timezone-field==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.5
lxml==5.1.0
oscrypto==1.3.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg==3.1.19
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyOpenSSL==24.0.0
pypdf==4.0.2
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
qrcode==7.4.2
redis==3.5.3
reportlab==4.0.9
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
Twisted==23.10.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.15
zope.interface==6.2
